adapt style, make play page work
