--- front/src/pages/PlayPage.css
@@ -1,54 +1,0 @@
-html,
-body {
-  margin: 0;
-  height: 100%;
-  background-color: rgb(31, 1, 1);
-}
-
-
-
-.buttons {
-  margin: 50px;
-  display: flex;
-  flex-wrap: wrap;
-}
-
-@media (max-width: 625px) {
-  .buttons {
-    flex-direction: column;
-  }
-}
-
-.buttons button {
-  flex-shrink: 0;
-  padding: 10px;
-  border-radius: 5px;
-  border: none;
-  color: #fff;
-  font-size: 18px;
-  transition: 200ms;
-  margin: 10px;
-  font-weight: bolder;
-  width: 160px;
-  box-shadow: 0px 0px 30px 0px rgba(0, 0, 0, 0.1);
-}
-
-.card-container {
-  position: relative;
-  width: 100%;
-  max-width: 550px;
-  height: 550px;
-  display: flex;
-  justify-content: center;
-  margin-bottom: 20px; /* Increased bottom margin */
-}
-
-
-.buttons button:hover {
-  transform: scale(1.05);
-}
-
-.MuiTypography-h6 {
-  color: aqua;
-}
-
