+++ front/src/pages/PlayPage.tsx
@@ -1,47 +1,54 @@
-import TinderCard from 'react-tinder-card';
-import type { Exercise } from '@shared_types';
-import { createRef, useEffect, useMemo, useRef, useState } from 'react';
+import {
+  type Email,
+  type Text as ExerciseText,
+  type Exercise,
+} from '@shared_types';
+import { useEffect, useRef, useState, CSSProperties } from 'react';
 import { useAppDispatch, useAppSelector } from '../redux/hook';
 import { setUser } from '../redux/slices';
 import ApiSdk from '../api/apiSdk';
-import { Button, Container, Typography } from '@mui/material';
-import './PlayPage.css';
 import FullScreenSpinner from '../components/FullScreenSpinner';
+import { Button, Typography } from '@mui/material';
+import EmailCard from '../components/EmailCard';
+import TextCard from '../components/TextCard';
+import './SwipeableCard.css';
+import PopUpFeedback from '../components/PopUpFeedback';
+import ScoreDisplay from '../components/ScoreDisplay';
 
-declare type Direction = 'left' | 'right' | 'up' | 'down';
-
-declare interface TinderCardAPI {
-  /**
-   * Programmatically trigger a swipe of the card in one of the valid directions `'left'`, `'right'`, `'up'` and `'down'`. This function, `swipe`, can be called on a reference of the TinderCard instance. Check the [example](https://github.com/3DJakob/react-tinder-card-demo/blob/master/src/examples/Advanced.js) code for more details on how to use this.
-   *
-   * @param dir The direction in which the card should be swiped. One of: `'left'`, `'right'`, `'up'` and `'down'`.
-   */
-  swipe(dir?: Direction): Promise<void>;
-
-  /**
-   * Restore swiped-card state. Use this function if you want to undo a swiped-card (e.g. you have a back button that shows last swiped card or you have a reset button. The promise is resolved once the card is returned
-   */
-  restoreCard(): Promise<void>;
+export interface PopUpResult {
+  isCorrect: boolean;
+  feedback: string;
 }
 
 function PlayPage() {
-  const [exercises, setExercises] = useState<Exercise[]>([]);
+  const [exercises, setExercises] = useState<(Email | ExerciseText)[]>([]);
   const { userInfo } = useAppSelector((state) => state.user);
   const [score, setScore] = useState<number>(userInfo.overallScore);
   const hasFetched = useRef(false);
   const api = new ApiSdk();
   const { token } = useAppSelector((state) => state.session);
-
+  const [isDragging, setIsDragging] = useState(false);
+  const [position, setPosition] = useState({ x: 0, y: 0 });
+  const [popupOpen, setPopupOpen] = useState<boolean>(false);
+  const [popupResult, setPopupResult] = useState<PopUpResult>({
+    isCorrect: false,
+    feedback: '',
+  });
+  const [isComplete, setIsComplete] = useState(false);
   const dispatch = useAppDispatch();
 
-  // Disable scrolling only for this specific page
-  useEffect(() => {
-    document.body.style.overflow = 'hidden';
+  const cardRef = useRef<HTMLDivElement>(null);
+  const startPosRef = useRef({ x: 0, y: 0 });
+  const SWIPE_THRESHOLD = 200; // pixels to trigger a swipe
 
-    return () => {
-      document.body.style.overflow = 'auto';
-    };
-  }, []);
+  // Disable scrolling only for this specific page
+  // useEffect(() => {
+  //   document.body.style.overflow = 'hidden';
+
+  //   return () => {
+  //     document.body.style.overflow = 'auto';
+  //   };
+  // }, []);
 
   useEffect(() => {
     if (!hasFetched.current) {
@@ -52,6 +59,14 @@ function PlayPage() {
     }
     // eslint-disable-next-line
   }, [token]);
+
+  const isEmail = (exercise: Exercise): exercise is Email => {
+    return exercise.type === 'email';
+  };
+
+  const isText = (exercise: Exercise): exercise is ExerciseText => {
+    return exercise.type === 'text';
+  };
 
   const [currentIndex, setCurrentIndex] = useState(exercises.length - 1);
   const currentIndexRef = useRef(currentIndex);
@@ -72,80 +87,167 @@ function PlayPage() {
     currentIndexRef.current = exercises.length - 1;
   }, [exercises]);
 
-  const updateCurrentIndex = (val: number) => {
-    setCurrentIndex(val);
-    currentIndexRef.current = val;
+  const handleMouseDown = (e: React.MouseEvent) => {
+    if (!cardRef.current) return;
+
+    setIsDragging(true);
+    startPosRef.current = {
+      x: e.clientX - position.x,
+      y: e.clientY - position.y,
+    };
   };
 
-  const swiped = (_direction: string, index: number) => {
-    updateCurrentIndex(index - 1);
+  const handleMouseMove = (e: React.MouseEvent) => {
+    if (!isDragging || !cardRef.current) return;
+
+    const newX = e.clientX - startPosRef.current.x;
+    const newY = e.clientY - startPosRef.current.y;
+
+    setPosition({ x: newX, y: newY });
   };
 
-  const outOfFrame = (idx: number) => {
-    console.log(
-      `card with index ${idx} left the screen!`,
-      currentIndexRef.current,
-    );
-    if (currentIndexRef.current >= idx && childRefs[idx].current) {
-      childRefs[idx].current?.restoreCard();
+  const handleMouseUp = () => {
+    if (!cardRef.current) return;
+
+    setIsDragging(false);
+
+    // Determine swipe direction based on position
+    if (Math.abs(position.x) > SWIPE_THRESHOLD) {
+      const direction = position.x > 0 ? 'right' : 'left';
+      handleSwipe(direction);
+    } else {
+      // Reset position if not swiped far enough
+      setPosition({ x: 0, y: 0 });
     }
   };
 
-  const swipe = async (dir: Direction) => {
-    if (canSwipe && currentIndex < exercises.length) {
-      const exercise = exercises[currentIndex];
-      const answeredAsScam = dir === 'left';
-      await childRefs[currentIndex].current?.swipe(dir); // Swipe the card!
-      if (exercise.scam === answeredAsScam) {
-        updateUserScore(score + 10);
-      }
+  const handleSwipe = (direction: 'left' | 'right') => {
+    if (currentIndex < 0) return;
+
+    const currentCard = exercises[currentIndex];
+
+    // Did they answer correctly?
+    const isCorrect =
+      (direction === 'left' && currentCard.scam) ||
+      (direction === 'right' && !currentCard.scam);
+
+    // Create popup response
+    setPopupResult({
+      isCorrect,
+      feedback: currentCard.feedback,
+    });
+
+    // Open popup
+    setPopupOpen(true);
+
+    // Remove swiped card
+    const newExercises = [...exercises];
+    newExercises.splice(currentIndex, 1);
+    setExercises(newExercises);
+
+    // Update current index
+    if (currentIndex === 0) {
+      setIsComplete(true);
+    } else {
+      setCurrentIndex((prev) => Math.max(prev - 1, -1));
+    }
+
+    // Reset position
+    setPosition({ x: 0, y: 0 });
+
+    // Update score if correct
+    if (isCorrect) {
+      updateUserScore(score + 10);
     }
   };
 
-  const childRefs = useMemo(
-    () =>
-      Array(exercises.length)
-        .fill(0)
-        .map(() => createRef<TinderCardAPI>()),
-    [exercises.length],
-  );
+  const handleClosePopup = () => {
+    setPopupOpen(false);
+    if (isComplete) {
+      setCurrentIndex((prev) => prev - 1);
+    }
+  };
 
-  const canSwipe = currentIndex >= 0;
+  // Styles for the card based on dragging state
+  const cardStyle: CSSProperties = {
+    transform: `translate(${position.x}px, ${position.y}px) rotate(${position.x / 20}deg)`,
+    transition: isDragging ? 'none' : 'transform 0.3s ease-out',
+  };
+
+  //If no cards left
+  if (currentIndex < 0) {
+    return <Typography variant='h6'>No more exercises!</Typography>;
+  }
+
+  // const currentCard = exercises[currentIndex]?.;
+  const currentExercise = exercises[currentIndex];
+
+  if (!currentExercise) {
+    return null; // Handle the case where currentExercise is undefined
+  }
 
   if (!exercises.length) return <FullScreenSpinner />;
 
   return (
-    <Container>
-      <Typography>Current score : {score}</Typography>
-      <Container className='cardContainer' data-testid='cardContainer'>
-        {exercises.map((card, index) => (
-          <TinderCard
-            onSwipe={(dir) => swiped(dir, index)}
-            onCardLeftScreen={() => outOfFrame(index)}
-            preventSwipe={['down', 'up']}
-            className='card'
-            ref={childRefs[index]}
-            key={card.message}
+    <div className='app-container'>
+      {/* <Typography className='score-display'>Current score: {score}</Typography> */}
+      <ScoreDisplay score={score} />
+      <div className='wrapper'>
+        <div className='card-container'>
+          <div
+            ref={cardRef}
+            className='swipeable-card'
+            style={cardStyle}
+            onMouseDown={handleMouseDown}
+            onMouseMove={handleMouseMove}
+            onMouseUp={handleMouseUp}
+            onMouseLeave={handleMouseUp}
           >
-            <Container className='innerCard'>{card.message}</Container>
-          </TinderCard>
-        ))}
-      </Container>
-      <Container className='buttons'>
-        <Button
-          style={{ backgroundColor: !canSwipe ? '#c3c4d3' : 'red' }}
-          onClick={() => swipe('left')}
-        >
-          Scam!
-        </Button>
-        <Button
-          style={{ backgroundColor: !canSwipe ? '#c3c4d3' : 'green' }}
-          onClick={() => swipe('right')}
-        >
-          Not a scam
-        </Button>
-      </Container>
-    </Container>
+            {isEmail(currentExercise) ? (
+              <EmailCard
+                sender={currentExercise.emailSender}
+                message={currentExercise.message}
+              />
+            ) : isText(currentExercise) ? (
+              <TextCard
+                sender={currentExercise.phoneNumber}
+                message={currentExercise.message}
+              />
+            ) : null}
+          </div>
+        </div>
+        <div className='button-container'>
+          <Button
+            variant='contained'
+            style={{
+              backgroundColor: currentIndex < 0 ? '#c3c4d3' : 'red',
+              minWidth: '100px',
+            }}
+            onClick={() => handleSwipe('left')}
+          >
+            Corrupt!
+          </Button>
+          <Button
+            variant='contained'
+            style={{
+              backgroundColor: currentIndex < 0 ? '#c3c4d3' : 'green',
+              minWidth: '100px',
+            }}
+            onClick={() => handleSwipe('right')}
+          >
+            Honorable
+          </Button>
+        </div>
+      </div>
+
+      {/* Result Popup */}
+      <PopUpFeedback
+        popupOpen={popupOpen}
+        handleClosePopup={handleClosePopup}
+        popupResult={popupResult}
+        isComplete={isComplete}
+      />
+    </div>
   );
 }
